Mini refonte Tortues et Trivial
- Mise en dossier
- Wait_for pour Tortues
- Optimisations
